Moved MapLegendLevel styled-components to MapLegendLevel/styles.js
