+++ src/components/MapLegendLevel/__tests__/MapLegendLevelStyledComponents.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { shallow } from "enzyme";
-import { LevelWrapper, LevelColor, LevelValue } from "../MapLegendLevel";
+import { LevelWrapper, LevelColor, LevelValue } from "../styles";
 
 describe("Tests for MapLegendLevel styled components", () => {
   test("renders LevelWrapper", () => {
