+++ src/components/MapLegendLevel/index.js
@@ -1,9 +1,5 @@
 import React from "react";
-import {
-  LevelWrapper,
-  LevelValue,
-  LevelColor,
-} from "../styled-components/MapLegendLevel";
+import { LevelWrapper, LevelValue, LevelColor } from "./styles";
 
 const MapLegendLevel = ({ color, value }) => {
   return (
